test(memory): split session visibility coverage
